+++ Assets/Scripe/Bean/Hero.cs
@@ -4,9 +4,9 @@ using System.Collections;
 public class Hero
 {
 	public long role_lv;
-	public float role_hp;
-	public float role_attack;
-	public float role_defense;
+	public double role_hp;
+	public double role_attack;
+	public double role_defense;
 	public string lvup_crystal;
 	public float attack_speed;
 	public float attack_range;
@@ -14,9 +14,9 @@ public class Hero
     public float hit;
     public float dod;
     public float cri;
-    public float cri_dam;
+    public double cri_dam;
     public float speed_up;
-    public float real_dam;
+    public double real_dam;
     public float range_type;
 
     public BigNumber mLvupCrystal;
