+++ Assets/Scripe/Bean/HurtStatus.cs
@@ -3,10 +3,10 @@ using System.Collections;
 
 public class HurtStatus
 {
-    public float blood;
+    public double blood;
     public bool isCrt;
     public bool isRate;
-    public HurtStatus(float blood, bool isCrt, bool isRate)
+    public HurtStatus(double blood, bool isCrt, bool isRate)
     {
         this.blood = blood;
         this.isCrt = isCrt;
